Moved the code to utils.py:   click_repl/__init__.py 	Splitted code into functions to test :   click_repl/utils.py 	modified:   tox.ini
