+++ tox.ini
@@ -2,7 +2,6 @@
 envlist =
     py{py,37,38,39,310}
     flake8
-    ; mypy
     click7
 
 minversion = 3.7.14
@@ -10,7 +9,7 @@ isolated_build = true
 
 [gh-actions]
 python =
-    3.7: py37, mypy, click7, flake8
+    3.7: py37, click7, flake8
     3.8: py38
     3.9: py39
     3.10: py310
@@ -26,11 +25,6 @@ deps =
     tox
 commands =
     pytest --basetemp={envtmpdir}
-
-; [testenv:mypy]
-; basepython = python3.7
-; deps = mypy
-; commands = mypy click_repl
 
 [testenv:flake8]
 basepython = python3.7
